VideoControls refactor + vignette
